refactor: move all parse and io methods from main to utils
